% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls}
\alias{calls}
\alias{gd_search}
\alias{gd_tags}
\alias{gd_sections}
\alias{gd_editions}
\alias{gd_items}
\title{Calls}
\usage{
gd_search(q = NULL, ..., pages = 1)

gd_tags(q = NULL, ..., pages = 1)

gd_sections(q = NULL, ..., pages = 1)

gd_editions(q = NULL, ..., pages = 1)

gd_items(items, ...)
}
\arguments{
\item{q}{The search query parameter supports \code{AND}, \code{OR} and \code{NOT} operators.}

\item{...}{Any other parameter, or filter, see the full list at \url{https://open-platform.theguardian.com/documentation/}.}

\item{pages}{Number of pages to collect.}

\item{items}{Vector of API links to items.}
}
\description{
All of The Guardian API endpoints.
}
\note{
This only "prepares" the API calls, use \code{\link{gd_call}} to execute them.
}
\examples{
\dontrun{
(to_search <- gd_search("debates", pages = 13))
results <- gd_call(to_search)

# select items to retrieve
items_to_get <- gd_items(results$apiUrl[1:13])
items <- gd_call(items_to_get)
}

}
