% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{gd_call}
\alias{gd_call}
\alias{gd_call.guardianCalls}
\title{Call}
\usage{
gd_call(..., batch_size = 12)

\method{gd_call}{guardianCalls}(..., batch_size = 12)
}
\arguments{
\item{...}{Objects of class \code{guardianCalls}.}

\item{batch_size}{Size of each batch.}
}
\description{
Executes calls from \code{guardianCalls} objects
}
\examples{
\dontrun{
(to_search <- gd_search("debates", pages = 13))
results <- gd_call(to_search)
}

}
