% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{gd_search}
\alias{gd_search}
\title{Search}
\usage{
gd_search(q = NULL, pages = 1)
}
\arguments{
\item{q}{The search query parameter supports \code{AND}, \code{OR} and \code{NOT} operators.}

\item{pages}{Number of pages to collect.}
}
\description{
The content endpoint returns all pieces of content in the API.
}
